requirements.txt
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
iniconfig==2.3.0
joblib==1.5.2
kiwisolver==1.4.9
llvmlite==0.45.1
matplotlib==3.10.7
numba==0.62.1
numpy==2.3.4
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
statsmodels==0.14.5
threadpoolctl==3.6.0
tzdata==2025.2
pyarrow>=17.0.0
tqdm==4.66.1
